Fix resyncing the ticket database after unexpected shutdown
The ticket database would restore from the genesis block if it failed
the check to see if the tip matched with the database best block. This
corrects the resyncing such that if it finds a matching block for the
ticket database top block in the database, it instead resyncs from that
block instead of the genesis block.