Check for hidden votes by ticket hash, not vote hash
A bug could cause the failure to create a template from parent
if a cached block template had a vote spending the same ticket
with a different vote hash as compared to the recently added
block. Now votes are checked by the tickets they spend rather
than the vote hashes.

Fixes #142.