Start RPC server earlier, add pipes for parent process IPC.
The RPC server is now the first subsystem that is started.  Because
the RPC server would not be usable without the rest of the subsystems
running, all requests are disabled until bringup has finished.  This
has the added benefit of erroring earlier if the RPC listeners can not
be created.

Rewrite startup/shutdown logic to simplify shutdown signaling.  All
cleanup is now run from deferred functions in the main function.

Add two new config options to set the read and write ends of a pair of
pipes.  This is used as a simple mechanism for a parent process to
communicate with, observe, and manage the lifetime of a child dcrd
process.  When the RX (read end) pipe is closed, clean shutdown
automatically begins.

Add a new flag --lifetimeevents to create and send lifetime event
notifications over the TX (write end) pipe during bringup and
shutdown.  This allows the parent process to observe which subsystems
are currently starting, running, and stopping.

Fixes #297.
Fixes #298.