Fix forced reorganizations
A bug in evaluating the correctness of a side chain block caused
forced reorganizations to fail post new database code. This was
corrected and children of the same parent at HEAD should now be
correctly evaluated and able to reorganize to one another without
issue. A test was also added for forced reorganizations.