Add vote tallying architecture
Another peripheral database to be used later in consensus has been
added.  This database tracks tallying and summation of votes from
stakeholder vote voteBits.  The intermediary tallies are cached and
the evaluation of tallies over long periods of time is highly
efficient.  Below is the benchmark for tallying the results of
voting over a 28-day period.

BenchmarkFindingVerdicts-8  200000  7955  ns/op  704 B/op  8 allocs/op

As with the new ticket stake nodes, the storage in the blockchain
block nodes is dynamic.  On startup, the cache of the tallies is
loaded along with the current tally for the best node.  Any
intermediary or sidechain tallies can be quickly generated by either
accessing the cache or the database.

A large degree of test infrastructure was introduced to best test
edge cases relating the the new vote tallying.  A dummy database was
added into dcrd/database so that very long sidechain evaluations of
tallies could be tested, better ensuring that long reorganizations
will not pose a risk in forking.  These tests use dynamic functions
to set voteBits and can easily be modified to further test the
infrastructure.

The new database buckets require an upgrade on startup.  Code for
this upgrade has been added, and the upgrade itself takes seconds on
mainnet and testnet.

Infrastructure for testing some of the blockchain API relating to
accessory stake data for block nodes was also added.  A few minor
bugs were discovered and patched.  A number of comments that were
inaccurate or mislabeled in the stake package were also fixed.
Benchmarks were moved to their own file, benchmark_test.go.