Handle stop blocks outside the main chain in locateBlocks()
Leave endIdx unbounded in locateBlocks() if the stop block is not on
the main chain. This causes as many blocks as possible to be stuffed
in the response, partially addressing #427.

XXX This is tentative at best. The locking needs verification, and the
whole logic behind locateBlocks() might need to change to accomodate
the other requirements lined out in #427.