rpcserver: Return RPC errors from block template.
This modifies the RPC server blockTemplateResult function to return RPC
errors as expected instead of using mempool and chain errors.

In addition to being more correct, it also helps pave the way for
mempool separation since its internal error funcs will no longer be
available to abuse as this code was doing.