blockchain: Remove impossible condition checks.
This removes a couple of checks that are completely impossible to hit.

These are impossible to hit because there is a maximum protocol limit of
1MiB and the absolute minimum possible transaction size that will
successfully deserialize with no inputs and no outputs is 15 bytes
(which is itself an invalid transaction anyways), which means a maximum
of 66,666 of these invalid transactions could fit into a 1MiB block.
Since the check is comparing against 393,216 on mainnet and 1,000,000 on
testnet and simnet, it can easily be seen it simply isn't possible to
hit.

Technically, if it were possible to hit these conditions, changing them
would be a hard fork, however, since they're impossible to hit, this is
effectively a noop change in regards to consensus.