connmgr: Add ability to remove pending connections
This is a port of btcd commit 548c0f499b24e31fa18fe6a48493e01175b54fe8.

This commit adds the ability for callers to remove pending connections
via a call to the Remove() method. With this change, upstream users of
this package can use the connmgr for more elaborate connectivity needs
as they can now cancel pending connections that are no longer needed.

The test was adapted to use the new DialAddr function, given the
difference in the function signature for Dial between btcd and dcrd.