rpcserver: fix empty ssgen txin[0] returned by get/searchrawtransaction(s)
Only Vin[0] was being set for stakebase/ssgen transactions in the vin lists created by createVinList and createVinListPrevOut.  Fix this to allow Vin[1] to be set too.
Also fix the Stakebase field of dcrjson.VinPrevOut and dcrjson.Vin being set incorrectly set to the zero hash instead of the actual hex-encoded signature script of the txin.