connmgr: Add cancellation of pending requests
This is a port of btcd commit b26daffac9dfe496c91a859c35a8bbe33c054b61.

 This commit extends the work started by roasbeef in the
 previous commit to bring full cancellation of pending
 connection requests. It also adds minor refactors to
 channel send/receives to help cleanup potentially
 lingering go routines.

The tests in the original commit were adapted to conform to dcrd's
function signatures and a few typos were fixed. The new state introduced
by the commit (ConnCanceled) was changed to be a new one at the end of
the possible states list and the ConnFailing state was renamed back to
ConnFailed. This was done to prevent having to perform a major version
bump in the connmgr module, given those states are part of the public
API.