stake: add ExpiredByBlock to stake.Node
This adds the (*Node).ExpiredByBlock to the stake
package. MissedByBlock includes both missed (a
winner with no vote) and expired. ExpiredByBlock
gives only tickets that expired.

Note that these functions check the undo data of
a stake.Node so they results reflect the latest
block to be connected and written to the backing
database.