peer: Remove superfluous dup version check.
This removes an impossible to hit duplicate version check and adds a
test to ensure the behavior is correct.

Previously, the check was necessary because the handler was a part of
the input handler loop, however, the initial negotiation logic was
refactored some time ago to be synchronous and thus it is no longer
possible to receive a second version message in the refactored code
path.