do not include revoke as expire
When a ticket initially expires, it is tagged as Missed and Expired,
then put into the missedTickets bucket.
Later when and if it is revoked, it is extracted from the missed tickets
bucket, retaining both the Missed and Expired flags, marked Revoked, and
put into the revokedTickets bucket.
When checking the undo data with ExpiredByBlock, it is thus necessary
to check (Expired && ! Revoked), otherwise the same ticket may be
reported as expired twice by different blocks.
The exact same issue with semantic affects MissedByBlock, just for both
expires and real misses.
A similar problem does NOT affect SpendByBlock as this applies only to
votes, which are just removed from the liveTickets bucket, but not
moved from bucket to bucket.