blockchain: Add BuildMerkleTreeStore alternative for MsgTx
A new function, BuildMsgTxMerkleTreeStore, is added.  It performs the
same operation as BuildMerkleTreeStore but uses a slice of wire.MsgTx
transactions as input rather than a slice of dcrutil.Tx wrapper types.

To avoid some code duplication, the construction of the merkle tree is
abstracted out to a helper function.