multi: Revert to using a single root module
The go1.11 module system works reasonably well when the dependencies are
housed in separate repositories, but creating multiple submodules within
the same repository just does not work well and significantly hinders
the ability to update the code.

For example, consider a change that needs to touch several modules, such
as a testnet reset.  Since each submodule is effectively version locked
the updates can't be made in a single PR since it would cause major
version bumps of the submodules, which in turn require new tags and
updates all of the import paths to include the new major version, but of
course it's not possible to tag the relevant commits until they've been
merged.

Further, it means that updates which fix bugs in the submodules would
not automatically be applied when building the latest master as the new
go1.11 MVS algorithm, as its name implies, selects the minimum version
per the existing go.mod file.  In other words, it would select the
versions of the modules without the fixes, until new submodule tags are
created and the root go.mod got updated to use the new version.

Essentially having multiple submodules in the repository is completely
untenable with the current iteration of the upcoming go1.11 module
system.

Consequently, this removes all of the submodules created in the
repository in anticipation of the go1.11 release in favor a single root
module which does not have any of the aforementioned issues.