[release-v1.3] blockchain: Mark fastadd block valid.
This modifies the block connection logic to also mark the status of
blocks that are not already known to be valid in the fast add code path
as valid.

This change is being made because the intent of the fast add path is to
skip expensive checks on blocks which are already assumed to be valid
due to checkpoints and therefore should be marked accordingly.

Also, since the status for those blocks in the block index does not have
them marked valid, mark them all ancestors of the current tip as valid
init time.