[release-v1.3] rpcserver: Cleanup cfilter handling.
This modifies the getcfilter and getcfilterheader RPC handlers to return
an error when an unknown block is provided as opposed an empty filter in
the case of the former and an all-zero hash in the case of the latter.

It also modifies the error returned when an unknown filter type is
specified to indicate an invalid parameter as opposed to a misc error so
it is more consistent with the other handlers.

Finally, it removes the debug logs in the error cases since the errors
are now returned to the caller.