blockmanager: handle txs in invalid blocks
Current code results in transactions from invalidated blocks being
removed from the mempool instead of remaining there to be mined in a
future block.

The new version ensures that transactions are *only* removed once they
are in a block voted as valid.

This only applies to regular transactions, since stake transactions are
not affected by the validity bits.