blockchain: Store interrupt channel with state.
This modifies the Blockchain struct to keep a reference to the interrupt
channel (if any) provided by the caller.  While currently it is not used
outside of the initialization path, having it available in the state
will be useful for other use cases such as being able to decouple reorgs
from the download state and allowing interrupts of long reorgs.