build: Support MacOS sed for obtaining module list.
This modifies the code in run_tests.sh which obtains the stripped list
of modules to test based on the output of the go list command to work
with MacOS as well.  It seems that sed on MacOS does not like the "or"
regexp syntax, so this modifies it to perform each branch separately.