chaincfg: Unexport internal errors.
These errors are only used internally and thus should not be exported.

While here, it also removes the completely unused VoteBitsNotFound
error.

It should be noted that this is a breaking change to the API and thus
will need a v2 major version bump of the chaincfg module to be published
before the changes can be externally consumed.