blockchain: Remove unused CheckWorklessBlockSanity.
This removes the CheckWorklessBlockSanity function since nothing uses
it anymore.  It was used for testing templates and unit tests in the
past, but the former is now handled by CheckConnectBlockTemplate and the
latter is handled by the full block tests.

It should be noted that this is a breaking change to the API and thus
will need a major version bump of the blockchain module to be published
before the next release.